+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+#### 1.1.0
+-   Added JsonxCodec, JsonxDecoder, and JsonxEncoder
+-	Added TypeMirror caching to improve the performance of `decode`
+
 #### 1.0.9
 -	Changed fixed length lists into growable lists in decoded objects
+
 #### 1.0.8
 -	Added support for encoding/decoding memmbers of superclasses (which
 	include mixins also)